speed_limit_system/requirements.txt
streamlit==1.40.2
streamlit-autorefresh==1.0.1

# Machine learning (NN training + inference)
scikit-learn==1.5.2
joblib==1.4.2

# LLM client (Grok via groq-python)
groq==0.13.0

# Environment variable management
python-dotenv==1.0.0

# Testing
pytest==8.3.3

# Data processing (for training data loading in Live mode)
pandas==2.2.3
numpy==1.26.4

# For typing (Python 3.10+ has everything built-in)
# typing-extensions==4.12.2   # Only needed if you use Python < 3.10